feat(notification): add simple fade out animation when it's closed
