+++ src/notification/keyframes.ts
@@ -1,0 +1,12 @@
+import { keyframes } from '@emotion/react';
+
+export function notificationExit() {
+  return keyframes`
+    from {
+      opacity: 1;
+    }
+    to {
+      opacity: 0;
+    }
+  `;
+}
